+++ R/revulytics_auth.R
@@ -27,13 +27,20 @@
 #' end_date <- Sys.Date() - 1
 #' session_id <- revulytics_auth(rev_user, rev_pwd)
 #' }
-
+ 
 revulytics_auth <- function(rev_username, rev_password) {
-  revulytics_login <- httr::POST("https://api.revulytics.com/auth/login",
+
+  revulytics_login <- httr::RETRY("POST",
+                                 url = "https://api.revulytics.com/auth/login",
                                  body = list(user = rev_username,
                                              password = rev_password,
                                              useCookies = FALSE),
-                                 encode = "json")
+                                 encode = "json",
+                                 times = 4,
+                                 pause_min = 10,
+                                 terminate_on = NULL,
+                                 terminate_on_success = TRUE,
+                                 pause_cap = 5)
   
   check_status(revulytics_login)